requirements/pi.txt
-r base.txt

RPi.GPIO
